[post creation page] Add post_creation_page.py page object and constants
